--- src/Selectors/index.js
@@ -1,1 +1,0 @@
-export { default, toggleSingle, toggleMultiple } from './BaseSelect'
