+++ src/Selects/index.js
@@ -1,0 +1,6 @@
+export {
+  default as BaseSelect,
+  toggleSingle,
+  toggleMultiple
+} from './BaseSelect'
+export { default as LabeledSelect } from './LabeledSelect'
