+++ src/Tabs/Tabs.module.scss
