+++ stories/CheckboxesStory.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
-import Checkboxes from '../src/Selectors/Checkboxes'
+import Checkboxes from '../src/Checkboxes'
 import Button from '../src/Button'
 import ColorModeComparison from './ColorModeComparison'
